deleted JSONConverter.swift and moved things into WeatherAPIClient
